+++ example/example.dart
@@ -3,10 +3,11 @@ import 'dart:async';
 import 'package:squadron/squadron.dart';
 
 import 'sample_worker.dart';
+import 'worker_monitor.dart';
 
 void main() async {
-  final loops = 10;
-  final max = 500;
+  final loops = 100;
+  final max = 50;
 
   print('');
   print('loops = $loops');
@@ -22,7 +23,7 @@ void main() async {
     final syncTs = DateTime.now().microsecondsSinceEpoch;
     for (var loop = 0; loop < loops; loop++) {
       final syncFutures = <Future>[];
-      for (var n = 0; n <= max; n++) {
+      for (var n = 0; n < max; n++) {
         syncFutures.add(Future.value(SampleWorker.cpuOperationImpl(n)));
         syncFutures.add(SampleWorker.ioOperationImpl(n));
       }
@@ -40,14 +41,18 @@ void main() async {
     final maxParallel = 2;
 
     pool = WorkerPool(() => SampleWorker(),
-        maxWorkers: maxWorkers, maxParallel: maxParallel);
+        minWorkers: 1, maxWorkers: maxWorkers, maxParallel: maxParallel);
     await pool.start();
+
+    final monitor = WorkerMonitor(pool, 100);
+    monitor.start();
 
     final asyncTs = DateTime.now().microsecondsSinceEpoch;
     for (var loop = 0; loop < loops; loop++) {
       final asyncFutures = <Future>[];
-      for (var n = 0; n <= max; n++) {
-        asyncFutures.add(pool.compute((worker) => worker.cpuOperation(n)));
+      for (var n = 0; n < max; n++) {
+        asyncFutures.add(
+            Future.value(pool.compute((worker) => worker.cpuOperation(n))));
         asyncFutures.add(pool.compute((worker) => worker.ioOperation(n)));
       }
       await Future.wait(asyncFutures);
@@ -55,9 +60,35 @@ void main() async {
     final asyncElapsed = DateTime.now().microsecondsSinceEpoch - asyncTs;
 
     print('async version completed in ${Duration(microseconds: asyncElapsed)}');
-    for (var stat in pool.stats) {
+
+    print('waiting for monitor to stop workers...');
+    final ts = DateTime.now().microsecondsSinceEpoch;
+    var prevSize = 0;
+    while (true) {
+      final size = pool.size;
+      if (prevSize != size) {
+        print('   pool.size = $size');
+        prevSize = size;
+      }
+      if (size == pool.minWorkers) break;
+      await Future.delayed(monitor.maxIdle ~/ 500);
+      if (DateTime.now().microsecondsSinceEpoch - ts >
+          monitor.maxIdle.inMicroseconds * 2) {
+        print('Houston, we\'ve got a problem...');
+      }
+    }
+
+    print('worker stats:');
+    for (var stat in pool.fullStats) {
       print('  * $stat');
     }
+
+    monitor.stop();
+
+    print('pool stats:');
+    print(
+        '  * load = ${pool.workload}, max load = ${pool.maxWorkload}, total load = ${pool.totalWorkload}, total errors = ${pool.totalErrors}');
+
     print('');
   } on WorkerException catch (e) {
     print(e.message);