+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.2
+
+- Added support for minimum count of workers in pool
+- Make sure Isolate is killed when the worker is stopped
+- Keep track of stats from stopped workers
+- Update tests & examples (proper error handling, command map, worker monitoring)
+
 ## 1.0.1
 
 - Improve error handling.